refactor: more forge crate modules
