+++ cli/src/cmd/forge/test.rs
@@ -15,11 +15,12 @@ use forge::{
     decode::decode_console_logs,
     executor::opts::EvmOpts,
     gas_report::GasReport,
+    result::{SuiteResult, TestKind, TestResult},
     trace::{
         identifier::{EtherscanIdentifier, LocalTraceIdentifier},
         CallTraceDecoderBuilder, TraceKind,
     },
-    MultiContractRunner, MultiContractRunnerBuilder, SuiteResult, TestFilter, TestKind,
+    MultiContractRunner, MultiContractRunnerBuilder, TestFilter,
 };
 use foundry_common::evm::EvmArgs;
 use foundry_config::{figment::Figment, Config};
@@ -333,7 +334,7 @@ pub struct Test {
     /// The signature of the solidity test
     pub signature: String,
     /// Result of the executed solidity test
-    pub result: forge::TestResult,
+    pub result: TestResult,
 }
 
 impl Test {
@@ -366,17 +367,17 @@ impl TestOutcome {
     }
 
     /// Iterator over all succeeding tests and their names
-    pub fn successes(&self) -> impl Iterator<Item = (&String, &forge::TestResult)> {
+    pub fn successes(&self) -> impl Iterator<Item = (&String, &TestResult)> {
         self.tests().filter(|(_, t)| t.success)
     }
 
     /// Iterator over all failing tests and their names
-    pub fn failures(&self) -> impl Iterator<Item = (&String, &forge::TestResult)> {
+    pub fn failures(&self) -> impl Iterator<Item = (&String, &TestResult)> {
         self.tests().filter(|(_, t)| !t.success)
     }
 
     /// Iterator over all tests and their names
-    pub fn tests(&self) -> impl Iterator<Item = (&String, &forge::TestResult)> {
+    pub fn tests(&self) -> impl Iterator<Item = (&String, &TestResult)> {
         self.results.values().flat_map(|SuiteResult { test_results, .. }| test_results.iter())
     }
 
@@ -433,7 +434,7 @@ impl TestOutcome {
     }
 }
 
-fn short_test_result(name: &str, result: &forge::TestResult) {
+fn short_test_result(name: &str, result: &TestResult) {
     let status = if result.success {
         Paint::green("[PASS]".to_string())
     } else {
