+++ evm/src/executor/mod.rs
@@ -27,7 +27,9 @@ pub use revm::db::DatabaseRef;
 pub use revm::Env;
 
 use self::inspector::{InspectorData, InspectorStackConfig};
-use crate::{debug::DebugArena, trace::CallTraceArena, CALLER};
+use crate::{
+    debug::DebugArena, executor::inspector::DEFAULT_CREATE2_DEPLOYER, trace::CallTraceArena, CALLER,
+};
 use bytes::Bytes;
 use ethers::{
     abi::{Abi, Detokenize, Tokenize},
@@ -199,6 +201,25 @@ where
         );
 
         Executor { db, env, inspector_config, gas_limit }
+    }
+
+    /// Creates the default CREATE2 Contract Deployer for local tests and scripts.
+    pub fn deploy_create2_deployer(&mut self) -> eyre::Result<()> {
+        let create2_deployer_account = self.db.basic(DEFAULT_CREATE2_DEPLOYER);
+
+        if create2_deployer_account.code.is_none() ||
+            create2_deployer_account.code.as_ref().unwrap().is_empty()
+        {
+            let creator = "0x3fAB184622Dc19b6109349B94811493BF2a45362".parse().unwrap();
+            self.set_balance(creator, U256::MAX);
+            self.deploy(
+                creator,
+                hex::decode("604580600e600039806000f350fe7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe03601600081602082378035828234f58015156039578182fd5b8082525050506014600cf3").expect("Could not decode create2 deployer init_code").into(),
+                U256::zero(),
+                None
+            )?;
+        }
+        Ok(())
     }
 
     /// Set the balance of an account.
