+++ forge/src/result.rs
@@ -1,0 +1,148 @@
+//! test outcomes
+
+use crate::Address;
+use ethers::prelude::Log;
+use foundry_evm::{
+    fuzz::{CounterExample, FuzzedCases},
+    trace::{CallTraceArena, TraceKind},
+};
+use serde::{Deserialize, Serialize};
+use std::{collections::BTreeMap, fmt, time::Duration};
+
+/// Results and duration for a set of tests included in the same test contract
+#[derive(Clone, Serialize)]
+pub struct SuiteResult {
+    /// Total duration of the test run for this block of tests
+    pub duration: Duration,
+    /// Individual test results. `test method name -> TestResult`
+    pub test_results: BTreeMap<String, TestResult>,
+    // Warnings
+    pub warnings: Vec<String>,
+}
+
+impl SuiteResult {
+    pub fn new(
+        duration: Duration,
+        test_results: BTreeMap<String, TestResult>,
+        warnings: Vec<String>,
+    ) -> Self {
+        Self { duration, test_results, warnings }
+    }
+
+    pub fn is_empty(&self) -> bool {
+        self.test_results.is_empty()
+    }
+
+    pub fn len(&self) -> usize {
+        self.test_results.len()
+    }
+}
+
+/// The result of an executed solidity test
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct TestResult {
+    /// Whether the test case was successful. This means that the transaction executed
+    /// properly, or that there was a revert and that the test was expected to fail
+    /// (prefixed with `testFail`)
+    pub success: bool,
+
+    /// If there was a revert, this field will be populated. Note that the test can
+    /// still be successful (i.e self.success == true) when it's expected to fail.
+    pub reason: Option<String>,
+
+    /// Minimal reproduction test case for failing fuzz tests
+    pub counterexample: Option<CounterExample>,
+
+    /// Any captured & parsed as strings logs along the test's execution which should
+    /// be printed to the user.
+    #[serde(skip)]
+    pub logs: Vec<Log>,
+
+    /// What kind of test this was
+    pub kind: TestKind,
+
+    /// Traces
+    pub traces: Vec<(TraceKind, CallTraceArena)>,
+
+    /// Labeled addresses
+    pub labeled_addresses: BTreeMap<Address, String>,
+}
+
+impl TestResult {
+    /// Returns `true` if this is the result of a fuzz test
+    pub fn is_fuzz(&self) -> bool {
+        matches!(self.kind, TestKind::Fuzz(_))
+    }
+}
+
+/// Used gas by a test
+#[derive(Debug, Clone, Eq, PartialEq)]
+pub enum TestKindGas {
+    Standard(u64),
+    Fuzz { runs: usize, mean: u64, median: u64 },
+}
+
+impl fmt::Display for TestKindGas {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            TestKindGas::Standard(gas) => {
+                write!(f, "(gas: {})", gas)
+            }
+            TestKindGas::Fuzz { runs, mean, median } => {
+                write!(f, "(runs: {}, μ: {}, ~: {})", runs, mean, median)
+            }
+        }
+    }
+}
+
+impl TestKindGas {
+    /// Returns the main gas value to compare against
+    pub fn gas(&self) -> u64 {
+        match self {
+            TestKindGas::Standard(gas) => *gas,
+            // We use the median for comparisons
+            TestKindGas::Fuzz { median, .. } => *median,
+        }
+    }
+}
+
+/// Various types of tests
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum TestKind {
+    /// A standard test that consists of calling the defined solidity function
+    ///
+    /// Holds the consumed gas
+    Standard(u64),
+    /// A solidity fuzz test, that stores all test cases
+    Fuzz(FuzzedCases),
+}
+
+impl TestKind {
+    /// The gas consumed by this test
+    pub fn gas_used(&self) -> TestKindGas {
+        match self {
+            TestKind::Standard(gas) => TestKindGas::Standard(*gas),
+            TestKind::Fuzz(fuzzed) => TestKindGas::Fuzz {
+                runs: fuzzed.cases().len(),
+                median: fuzzed.median_gas(false),
+                mean: fuzzed.mean_gas(false),
+            },
+        }
+    }
+}
+
+#[derive(Clone, Debug, Default)]
+pub struct TestSetup {
+    /// The address at which the test contract was deployed
+    pub address: Address,
+    /// The logs emitted during setup
+    pub logs: Vec<Log>,
+    /// Call traces of the setup
+    pub traces: Vec<(TraceKind, CallTraceArena)>,
+    /// Addresses labeled during setup
+    pub labeled_addresses: BTreeMap<Address, String>,
+    /// Whether the setup failed
+    pub setup_failed: bool,
+    /// The reason the setup failed
+    pub reason: Option<String>,
+}
