+++ forge/src/traits.rs
@@ -1,0 +1,6 @@
+/// Extension trait for matching tests
+pub trait TestFilter: Send + Sync {
+    fn matches_test(&self, test_name: impl AsRef<str>) -> bool;
+    fn matches_contract(&self, contract_name: impl AsRef<str>) -> bool;
+    fn matches_path(&self, path: impl AsRef<str>) -> bool;
+}
